finish moving core-only deps generation events (M) to core
