+++ core/dbt/clients/git.py
@@ -3,7 +3,7 @@ import os.path
 
 from dbt.common.clients.system import run_cmd, rmdir
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import (
+from dbt.events.types import (
     GitSparseCheckoutSubdirectory,
     GitProgressCheckoutRevision,
     GitProgressUpdatingExistingDependency,
