+++ core/dbt/clients/registry.py
@@ -2,7 +2,7 @@ import functools
 from typing import Any, Dict, List
 import requests
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import (
+from dbt.events.types import (
     RegistryProgressGETRequest,
     RegistryProgressGETResponse,
     RegistryIndexProgressGETRequest,
