+++ core/dbt/common/events/types.proto
@@ -66,180 +66,6 @@ message GenericMessage {
 
 // M - Deps generation
 
-// M001
-message GitSparseCheckoutSubdirectory {
-    string subdir = 1;
-}
-
-message GitSparseCheckoutSubdirectoryMsg {
-    EventInfo info = 1;
-    GitSparseCheckoutSubdirectory data = 2;
-}
-
-// M002
-message GitProgressCheckoutRevision {
-    string revision = 1;
-}
-
-message GitProgressCheckoutRevisionMsg {
-    EventInfo info = 1;
-    GitProgressCheckoutRevision data = 2;
-}
-
-// M003
-message GitProgressUpdatingExistingDependency {
-    string dir = 1;
-}
-
-message GitProgressUpdatingExistingDependencyMsg {
-    EventInfo info = 1;
-    GitProgressUpdatingExistingDependency data = 2;
-}
-
-// M004
-message GitProgressPullingNewDependency {
-    string dir = 1;
-}
-
-message GitProgressPullingNewDependencyMsg {
-    EventInfo info = 1;
-    GitProgressPullingNewDependency data = 2;
-}
-
-// M005
-message GitNothingToDo {
-    string sha = 1;
-}
-
-message GitNothingToDoMsg {
-    EventInfo info = 1;
-    GitNothingToDo data = 2;
-}
-
-// M006
-message GitProgressUpdatedCheckoutRange {
-    string start_sha = 1;
-    string end_sha = 2;
-}
-
-message GitProgressUpdatedCheckoutRangeMsg {
-    EventInfo info = 1;
-    GitProgressUpdatedCheckoutRange data = 2;
-}
-
-// M007
-message GitProgressCheckedOutAt {
-    string end_sha = 1;
-}
-
-message GitProgressCheckedOutAtMsg {
-    EventInfo info = 1;
-    GitProgressCheckedOutAt data = 2;
-}
-
-// M008
-message RegistryProgressGETRequest {
-    string url = 1;
-}
-
-message RegistryProgressGETRequestMsg {
-    EventInfo info = 1;
-    RegistryProgressGETRequest data = 2;
-}
-
-// M009
-message RegistryProgressGETResponse {
-    string url = 1;
-    int32 resp_code = 2;
-}
-
-message RegistryProgressGETResponseMsg {
-    EventInfo info = 1;
-    RegistryProgressGETResponse data = 2;
-}
-
-// M010
-message SelectorReportInvalidSelector {
-    string valid_selectors = 1;
-    string spec_method = 2;
-    string raw_spec = 3;
-}
-
-message SelectorReportInvalidSelectorMsg {
-    EventInfo info = 1;
-    SelectorReportInvalidSelector data = 2;
-}
-
-// Skipped M011 and M012
-
-// M013
-message DepsNoPackagesFound {
-}
-
-message DepsNoPackagesFoundMsg {
-    EventInfo info = 1;
-    DepsNoPackagesFound data = 2;
-}
-
-// M014
-message DepsStartPackageInstall {
-    string package_name = 1;
-}
-
-message DepsStartPackageInstallMsg {
-    EventInfo info = 1;
-    DepsStartPackageInstall data = 2;
-}
-
-// M015
-message DepsInstallInfo {
-    string version_name = 1;
-}
-
-message DepsInstallInfoMsg {
-    EventInfo info = 1;
-    DepsInstallInfo data = 2;
-}
-
-// M016
-message DepsUpdateAvailable {
-    string version_latest = 1;
-}
-
-message DepsUpdateAvailableMsg {
-    EventInfo info = 1;
-    DepsUpdateAvailable data = 2;
-}
-
-// M017
-message DepsUpToDate {
-}
-
-message DepsUpToDateMsg {
-    EventInfo info = 1;
-    DepsUpToDate data = 2;
-}
-
-// M018
-message DepsListSubdirectory {
-    string subdirectory = 1;
-}
-
-message DepsListSubdirectoryMsg {
-    EventInfo info = 1;
-    DepsListSubdirectory data = 2;
-}
-
-// M019
-message DepsNotifyUpdatesAvailable {
-    repeated string packages = 1;
-}
-
-message DepsNotifyUpdatesAvailableMsg {
-    EventInfo info = 1;
-    DepsNotifyUpdatesAvailable data = 2;
-}
-
 // M020
 message RetryExternalCall {
     int32 attempt = 1;
@@ -259,140 +85,6 @@ message RecordRetryException {
 message RecordRetryExceptionMsg {
     EventInfo info = 1;
     RecordRetryException data = 2;
-}
-
-// M022
-message RegistryIndexProgressGETRequest {
-    string url = 1;
-}
-
-message RegistryIndexProgressGETRequestMsg {
-    EventInfo info = 1;
-    RegistryIndexProgressGETRequest data = 2;
-}
-
-// M023
-message RegistryIndexProgressGETResponse {
-    string url = 1;
-    int32 resp_code = 2;
-}
-
-message RegistryIndexProgressGETResponseMsg {
-    EventInfo info = 1;
-    RegistryIndexProgressGETResponse data = 2;
-}
-
-// M024
-message RegistryResponseUnexpectedType {
-    string response = 1;
-}
-
-message RegistryResponseUnexpectedTypeMsg {
-    EventInfo info = 1;
-    RegistryResponseUnexpectedType data = 2;
-}
-
-// M025
-message RegistryResponseMissingTopKeys {
-    string response = 1;
-}
-
-message RegistryResponseMissingTopKeysMsg {
-    EventInfo info = 1;
-    RegistryResponseMissingTopKeys data = 2;
-}
-
-// M026
-message RegistryResponseMissingNestedKeys {
-    string response = 1;
-}
-
-message RegistryResponseMissingNestedKeysMsg {
-    EventInfo info = 1;
-    RegistryResponseMissingNestedKeys data = 2;
-}
-
-// m027
-message RegistryResponseExtraNestedKeys {
-    string response = 1;
-}
-
-message RegistryResponseExtraNestedKeysMsg {
-    EventInfo info = 1;
-    RegistryResponseExtraNestedKeys data = 2;
-}
-
-// M028
-message DepsSetDownloadDirectory {
-    string path = 1;
-}
-
-message DepsSetDownloadDirectoryMsg {
-    EventInfo info = 1;
-    DepsSetDownloadDirectory data = 2;
-}
-
-// M029
-message DepsUnpinned {
-    string revision = 1;
-    string git = 2;
-}
-
-message DepsUnpinnedMsg {
-    EventInfo info = 1;
-    DepsUnpinned data = 2;
-}
-
-// M030
-message NoNodesForSelectionCriteria {
-    string spec_raw = 1;
-}
-
-message NoNodesForSelectionCriteriaMsg {
-    EventInfo info = 1;
-    NoNodesForSelectionCriteria data = 2;
-}
-
-// M031
-message DepsLockUpdating{
-    string lock_filepath = 1;
-}
-
-message DepsLockUpdatingMsg{
-    EventInfo info = 1;
-    DepsLockUpdating data = 2;
-}
-
-// M032
-message DepsAddPackage{
-    string package_name = 1;
-    string version = 2;
-    string packages_filepath = 3;
-}
-
-message DepsAddPackageMsg{
-    EventInfo info = 1;
-    DepsAddPackage data = 2;
-}
-
-//M033
-message DepsFoundDuplicatePackage{
-    map<string, string> removed_package = 1;
-}
-
-message DepsFoundDuplicatePackageMsg{
-    EventInfo info = 1;
-    DepsFoundDuplicatePackage data = 2;
-}
-
-//M034
-message DepsVersionMissing{
-    string source = 1;
-}
-
-message DepsVersionMissingMsg{
-    EventInfo info = 1;
-    DepsVersionMissing data = 2;
 }
 
 // Q - Node execution
