+++ core/dbt/common/events/types.py
@@ -1,6 +1,5 @@
 import json
 
-from dbt.constants import PIN_PACKAGE_URL
 from dbt.common.events.base_types import (
     DynamicLevel,
     DebugLevel,
@@ -48,146 +47,6 @@ from dbt.common.ui import red, green, yellow
 # =======================================================
 
 
-class GitSparseCheckoutSubdirectory(DebugLevel):
-    def code(self) -> str:
-        return "M001"
-
-    def message(self) -> str:
-        return f"Subdirectory specified: {self.subdir}, using sparse checkout."
-
-
-class GitProgressCheckoutRevision(DebugLevel):
-    def code(self) -> str:
-        return "M002"
-
-    def message(self) -> str:
-        return f"Checking out revision {self.revision}."
-
-
-class GitProgressUpdatingExistingDependency(DebugLevel):
-    def code(self) -> str:
-        return "M003"
-
-    def message(self) -> str:
-        return f"Updating existing dependency {self.dir}."
-
-
-class GitProgressPullingNewDependency(DebugLevel):
-    def code(self) -> str:
-        return "M004"
-
-    def message(self) -> str:
-        return f"Pulling new dependency {self.dir}."
-
-
-class GitNothingToDo(DebugLevel):
-    def code(self) -> str:
-        return "M005"
-
-    def message(self) -> str:
-        return f"Already at {self.sha}, nothing to do."
-
-
-class GitProgressUpdatedCheckoutRange(DebugLevel):
-    def code(self) -> str:
-        return "M006"
-
-    def message(self) -> str:
-        return f"Updated checkout from {self.start_sha} to {self.end_sha}."
-
-
-class GitProgressCheckedOutAt(DebugLevel):
-    def code(self) -> str:
-        return "M007"
-
-    def message(self) -> str:
-        return f"Checked out at {self.end_sha}."
-
-
-class RegistryProgressGETRequest(DebugLevel):
-    def code(self) -> str:
-        return "M008"
-
-    def message(self) -> str:
-        return f"Making package registry request: GET {self.url}"
-
-
-class RegistryProgressGETResponse(DebugLevel):
-    def code(self) -> str:
-        return "M009"
-
-    def message(self) -> str:
-        return f"Response from registry: GET {self.url} {self.resp_code}"
-
-
-class SelectorReportInvalidSelector(InfoLevel):
-    def code(self) -> str:
-        return "M010"
-
-    def message(self) -> str:
-        return (
-            f"The '{self.spec_method}' selector specified in {self.raw_spec} is "
-            f"invalid. Must be one of [{self.valid_selectors}]"
-        )
-
-
-class DepsNoPackagesFound(InfoLevel):
-    def code(self) -> str:
-        return "M013"
-
-    def message(self) -> str:
-        return "Warning: No packages were found in packages.yml"
-
-
-class DepsStartPackageInstall(InfoLevel):
-    def code(self) -> str:
-        return "M014"
-
-    def message(self) -> str:
-        return f"Installing {self.package_name}"
-
-
-class DepsInstallInfo(InfoLevel):
-    def code(self) -> str:
-        return "M015"
-
-    def message(self) -> str:
-        return f"Installed from {self.version_name}"
-
-
-class DepsUpdateAvailable(InfoLevel):
-    def code(self) -> str:
-        return "M016"
-
-    def message(self) -> str:
-        return f"Updated version available: {self.version_latest}"
-
-
-class DepsUpToDate(InfoLevel):
-    def code(self) -> str:
-        return "M017"
-
-    def message(self) -> str:
-        return "Up to date!"
-
-
-class DepsListSubdirectory(InfoLevel):
-    def code(self) -> str:
-        return "M018"
-
-    def message(self) -> str:
-        return f"and subdirectory {self.subdirectory}"
-
-
-class DepsNotifyUpdatesAvailable(InfoLevel):
-    def code(self) -> str:
-        return "M019"
-
-    def message(self) -> str:
-        return f"Updates available for packages: {self.packages} \
-                \nUpdate your versions in packages.yml, then run dbt deps"
-
-
 class RetryExternalCall(DebugLevel):
     def code(self) -> str:
         return "M020"
@@ -202,116 +61,6 @@ class RecordRetryException(DebugLevel):
 
     def message(self) -> str:
         return f"External call exception: {self.exc}"
-
-
-class RegistryIndexProgressGETRequest(DebugLevel):
-    def code(self) -> str:
-        return "M022"
-
-    def message(self) -> str:
-        return f"Making package index registry request: GET {self.url}"
-
-
-class RegistryIndexProgressGETResponse(DebugLevel):
-    def code(self) -> str:
-        return "M023"
-
-    def message(self) -> str:
-        return f"Response from registry index: GET {self.url} {self.resp_code}"
-
-
-class RegistryResponseUnexpectedType(DebugLevel):
-    def code(self) -> str:
-        return "M024"
-
-    def message(self) -> str:
-        return f"Response was None: {self.response}"
-
-
-class RegistryResponseMissingTopKeys(DebugLevel):
-    def code(self) -> str:
-        return "M025"
-
-    def message(self) -> str:
-        # expected/actual keys logged in exception
-        return f"Response missing top level keys: {self.response}"
-
-
-class RegistryResponseMissingNestedKeys(DebugLevel):
-    def code(self) -> str:
-        return "M026"
-
-    def message(self) -> str:
-        # expected/actual keys logged in exception
-        return f"Response missing nested keys: {self.response}"
-
-
-class RegistryResponseExtraNestedKeys(DebugLevel):
-    def code(self) -> str:
-        return "M027"
-
-    def message(self) -> str:
-        # expected/actual keys logged in exception
-        return f"Response contained inconsistent keys: {self.response}"
-
-
-class DepsSetDownloadDirectory(DebugLevel):
-    def code(self) -> str:
-        return "M028"
-
-    def message(self) -> str:
-        return f"Set downloads directory='{self.path}'"
-
-
-class DepsUnpinned(WarnLevel):
-    def code(self) -> str:
-        return "M029"
-
-    def message(self) -> str:
-        if self.revision == "HEAD":
-            unpinned_msg = "not pinned, using HEAD (default branch)"
-        elif self.revision in ("main", "master"):
-            unpinned_msg = f'pinned to the "{self.revision}" branch'
-        else:
-            unpinned_msg = None
-
-        msg = (
-            f'The git package "{self.git}" \n\tis {unpinned_msg}.\n\tThis can introduce '
-            f"breaking changes into your project without warning!\n\nSee {PIN_PACKAGE_URL}"
-        )
-        return yellow(f"WARNING: {msg}")
-
-
-class NoNodesForSelectionCriteria(WarnLevel):
-    def code(self) -> str:
-        return "M030"
-
-    def message(self) -> str:
-        return f"The selection criterion '{self.spec_raw}' does not match any nodes"
-
-
-class DepsLockUpdating(InfoLevel):
-    def code(self):
-        return "M031"
-
-    def message(self) -> str:
-        return f"Updating lock file in file path: {self.lock_filepath}"
-
-
-class DepsAddPackage(InfoLevel):
-    def code(self):
-        return "M032"
-
-    def message(self) -> str:
-        return f"Added new package {self.package_name}@{self.version} to {self.packages_filepath}"
-
-
-class DepsFoundDuplicatePackage(InfoLevel):
-    def code(self):
-        return "M033"
-
-    def message(self) -> str:
-        return f"Found duplicate package in packages.yml, removing: {self.removed_package}"
 
 
 # =======================================================
