+++ core/dbt/graph/selector.py
@@ -6,7 +6,7 @@ from .selector_methods import MethodManager
 from .selector_spec import SelectionCriteria, SelectionSpec, IndirectSelection
 
 from dbt.common.events.functions import fire_event, warn_or_error
-from dbt.common.events.types import SelectorReportInvalidSelector, NoNodesForSelectionCriteria
+from dbt.events.types import SelectorReportInvalidSelector, NoNodesForSelectionCriteria
 from dbt.node_types import NodeType
 from dbt.exceptions import (
     DbtInternalError,
