+++ core/dbt/task/deps.py
@@ -18,6 +18,9 @@ from dbt.contracts.project import Package
 
 from dbt.common.events.functions import fire_event
 from dbt.common.events.types import (
+    Formatting,
+)
+from dbt.events.types import (
     DepsAddPackage,
     DepsFoundDuplicatePackage,
     DepsInstallInfo,
@@ -28,7 +31,6 @@ from dbt.common.events.types import (
     DepsStartPackageInstall,
     DepsUpdateAvailable,
     DepsUpToDate,
-    Formatting,
 )
 from dbt.common.clients import system
