+++ tests/unit/test_events.py
@@ -281,36 +281,38 @@ sample_values = [
     core_types.FreshnessConfigProblem(msg=""),
     core_types.SemanticValidationFailure(msg=""),
     # M - Deps generation ======================
-    types.GitSparseCheckoutSubdirectory(subdir=""),
-    types.GitProgressCheckoutRevision(revision=""),
-    types.GitProgressUpdatingExistingDependency(dir=""),
-    types.GitProgressPullingNewDependency(dir=""),
-    types.GitNothingToDo(sha=""),
-    types.GitProgressUpdatedCheckoutRange(start_sha="", end_sha=""),
-    types.GitProgressCheckedOutAt(end_sha=""),
-    types.RegistryProgressGETRequest(url=""),
-    types.RegistryProgressGETResponse(url="", resp_code=1234),
-    types.SelectorReportInvalidSelector(valid_selectors="", spec_method="", raw_spec=""),
-    types.DepsNoPackagesFound(),
-    types.DepsStartPackageInstall(package_name=""),
-    types.DepsInstallInfo(version_name=""),
-    types.DepsUpdateAvailable(version_latest=""),
-    types.DepsUpToDate(),
-    types.DepsListSubdirectory(subdirectory=""),
-    types.DepsNotifyUpdatesAvailable(packages=["my_pkg", "other_pkg"]),
+    core_types.GitSparseCheckoutSubdirectory(subdir=""),
+    core_types.GitProgressCheckoutRevision(revision=""),
+    core_types.GitProgressUpdatingExistingDependency(dir=""),
+    core_types.GitProgressPullingNewDependency(dir=""),
+    core_types.GitNothingToDo(sha=""),
+    core_types.GitProgressUpdatedCheckoutRange(start_sha="", end_sha=""),
+    core_types.GitProgressCheckedOutAt(end_sha=""),
+    core_types.RegistryProgressGETRequest(url=""),
+    core_types.RegistryProgressGETResponse(url="", resp_code=1234),
+    core_types.SelectorReportInvalidSelector(valid_selectors="", spec_method="", raw_spec=""),
+    core_types.DepsNoPackagesFound(),
+    core_types.DepsStartPackageInstall(package_name=""),
+    core_types.DepsInstallInfo(version_name=""),
+    core_types.DepsUpdateAvailable(version_latest=""),
+    core_types.DepsUpToDate(),
+    core_types.DepsListSubdirectory(subdirectory=""),
+    core_types.DepsNotifyUpdatesAvailable(packages=["my_pkg", "other_pkg"]),
     types.RetryExternalCall(attempt=0, max=0),
     types.RecordRetryException(exc=""),
-    types.RegistryIndexProgressGETRequest(url=""),
-    types.RegistryIndexProgressGETResponse(url="", resp_code=1234),
-    types.RegistryResponseUnexpectedType(response=""),
-    types.RegistryResponseMissingTopKeys(response=""),
-    types.RegistryResponseMissingNestedKeys(response=""),
-    types.RegistryResponseExtraNestedKeys(response=""),
-    types.DepsSetDownloadDirectory(path=""),
-    types.DepsLockUpdating(lock_filepath=""),
-    types.DepsAddPackage(package_name="", version="", packages_filepath=""),
-    types.DepsFoundDuplicatePackage(removed_package={}),
+    core_types.RegistryIndexProgressGETRequest(url=""),
+    core_types.RegistryIndexProgressGETResponse(url="", resp_code=1234),
+    core_types.RegistryResponseUnexpectedType(response=""),
+    core_types.RegistryResponseMissingTopKeys(response=""),
+    core_types.RegistryResponseMissingNestedKeys(response=""),
+    core_types.RegistryResponseExtraNestedKeys(response=""),
+    core_types.DepsSetDownloadDirectory(path=""),
+    core_types.DepsLockUpdating(lock_filepath=""),
+    core_types.DepsAddPackage(package_name="", version="", packages_filepath=""),
+    core_types.DepsFoundDuplicatePackage(removed_package={}),
     core_types.DepsScrubbedPackageName(package_name=""),
+    core_types.DepsUnpinned(revision="", git=""),
+    core_types.NoNodesForSelectionCriteria(spec_raw=""),
     # Q - Node execution ======================
     types.RunningOperationCaughtError(exc=""),
     types.CompileComplete(),
@@ -387,8 +389,6 @@ sample_values = [
     types.RunningOperationUncaughtError(exc=""),
     types.EndRunResult(),
     types.NoNodesSelected(),
-    types.DepsUnpinned(revision="", git=""),
-    types.NoNodesForSelectionCriteria(spec_raw=""),
     types.CommandCompleted(
         command="",
         success=True,
